Finished with exception handling

+++ MoneyTransferApplication/src/main/java/com/EM2/MoneyTransferApplication/daoImpl/AccountDaoImpl.java
@@ -1,7 +1,6 @@
 package com.EM2.MoneyTransferApplication.daoImpl;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -12,12 +11,16 @@ import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.RowMapper;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataAccessResourceFailureException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.jdbc.CannotGetJdbcConnectionException;
 import org.springframework.jdbc.core.support.JdbcDaoSupport;
 import org.springframework.stereotype.Repository;
 
 import com.EM2.MoneyTransferApplication.dao.AccountDao;
 import com.EM2.MoneyTransferApplication.model.Account;
+import com.EM2.MoneyTransferApplication.rowmap.AccountRowMapper;
 
 @Repository
 public class AccountDaoImpl extends JdbcDaoSupport implements AccountDao{
@@ -36,18 +39,34 @@ public class AccountDaoImpl extends JdbcDaoSupport implements AccountDao{
 	public void createAccount(int customerId) {
 		logger.info("Creating a new account for customer: "+customerId);
 		String createAccQuery = "INSERT INTO account(user_id) VALUES (?);";
-		getJdbcTemplate().update(createAccQuery, new Object[] {
-				customerId
-			});
+		try {
+			getJdbcTemplate().update(createAccQuery, new Object[] {
+					customerId
+				});
+		} catch (CannotGetJdbcConnectionException e) {
+			logger.error("JDBCConnectionError inserting new account");
+			e.printStackTrace();
+		}catch(DataAccessException e) {
+			logger.error("DataAccessEx while inserting new account");
+			e.printStackTrace();
+		}
 	}
 
 	@Override
 	public void depositMoney(int accountId, double moneyAmount) {
 		logger.info("Depositing money to account: "+accountId);
 		String depositMoneyQuery= "UPDATE account SET balance=balance+? WHERE account_id=?";
-		getJdbcTemplate().update(depositMoneyQuery, new Object[] {
-				moneyAmount,accountId
-		});
+		try {
+			getJdbcTemplate().update(depositMoneyQuery, new Object[] {
+					moneyAmount,accountId
+			});
+		} catch (CannotGetJdbcConnectionException e) {
+			logger.error("JDBCConnectionError while depositing money");
+			e.printStackTrace();
+		}catch(DataAccessException e) {
+			logger.error("DataAccessEx while depositing money");
+			e.printStackTrace();
+		}
 	}
 
 	@Override
@@ -55,9 +74,17 @@ public class AccountDaoImpl extends JdbcDaoSupport implements AccountDao{
 		if(moneyAmount<=sourceAccount.getBalance()) {
 			logger.info("Transfering money from account:"+sourceAccount+" to account:"+destinationAccId);
 			String reduceBalanceQuery="UPDATE account SET balance=balance-? WHERE user_id=? AND account_id=?";
-			getJdbcTemplate().update(reduceBalanceQuery,new Object[] {
-				moneyAmount,sourceAccount.getUserId(),sourceAccount.getAccId()	
-			});
+			try {
+				getJdbcTemplate().update(reduceBalanceQuery,new Object[] {
+					moneyAmount,sourceAccount.getUserId(),sourceAccount.getAccId()	
+				});
+			} catch (CannotGetJdbcConnectionException e) {
+				logger.error("JDBCConnectionError while transfering money");
+				e.printStackTrace();
+			}catch(DataAccessException e) {
+				logger.error("DataAccessEx while traansfering money");
+				e.printStackTrace();
+			}
 			depositMoney(destinationAccId, moneyAmount);
 		}else {
 			throw new RuntimeException("You can't transfer more money than you have");
@@ -68,7 +95,16 @@ public class AccountDaoImpl extends JdbcDaoSupport implements AccountDao{
 	public List<Account> getAllAccounts() {
 		logger.info("Getting all accounts from database");
 		String allAccountsQuery="SELECT * FROM account";
-		List<Map<String, Object>> rows = getJdbcTemplate().queryForList(allAccountsQuery);
+		List<Map<String, Object>> rows=null;
+		try {
+			rows = getJdbcTemplate().queryForList(allAccountsQuery);
+		} catch (CannotGetJdbcConnectionException e) {
+			logger.error("JDBCConnectionError while getting all accounts");
+			e.printStackTrace();
+		}catch(DataAccessException e) {
+			logger.error("DataAccessEx while getting all accounts");
+			e.printStackTrace();
+		}
 		return createAccountList(rows);
 	}
 
@@ -76,7 +112,16 @@ public class AccountDaoImpl extends JdbcDaoSupport implements AccountDao{
 	public List<Account> getAccountsByCustomerId(int customerId) {
 		logger.info("Getting all accounts of customer: "+customerId+" from database");
 		String customerAccountsQuery="SELECT * FROM account WHERE user_id="+customerId;
-		List<Map<String, Object>> rows = getJdbcTemplate().queryForList(customerAccountsQuery);
+		List<Map<String, Object>> rows=null;
+		try {
+			rows=getJdbcTemplate().queryForList(customerAccountsQuery);
+		} catch (CannotGetJdbcConnectionException e) {
+			logger.error("JDBCConnectionError while getting customer accounts");
+			e.printStackTrace();
+		}catch(DataAccessException e) {
+			logger.error("DataAccessEx while getting customer accounts");
+			e.printStackTrace();
+		}
 		return createAccountList(rows);
 	}
 
@@ -97,25 +142,50 @@ public class AccountDaoImpl extends JdbcDaoSupport implements AccountDao{
 	public Account getAccountById(int accountId) {
 		logger.info("Getting account from account id:"+accountId);
 		String getAccountQuery = "SELECT * FROM account WHERE account_id=?;";
-		return getJdbcTemplate().queryForObject(getAccountQuery, new Object[] {accountId}, new RowMapper<Account>() {
-
-			@Override
-			public Account mapRow(ResultSet row, int rowNumber) throws SQLException {
-				int userId = (int) row.getInt("user_id");
-				int accId = (int) row.getInt("account_id");
-				double balance = (double) row.getDouble("balance");
-				
-				Account account = new Account(accId, balance, userId);
-				return account;
-			}
-		});
+		Account account=null;
+		try {
+			account=getJdbcTemplate().queryForObject(getAccountQuery, new Object[] {accountId}, new AccountRowMapper());
+		} catch (CannotGetJdbcConnectionException e) {
+			logger.error("JDBCConnectionError while getting account");
+			e.printStackTrace();
+		}catch(DataAccessException e) {
+			logger.error("DataAccessEx while getting account");
+			e.printStackTrace();
+		}
+		return account;
+		
 	}
 
 	@Override
-	public void deleteAccount(int accId) {
-		logger.info("Deleting account: "+accId+ "from database");
-		String deleteQuery = "DELETE FROM account WHERE account_id=?";
-		getJdbcTemplate().update(deleteQuery,accId);
+	public void deleteAccount(int accId,int customerId) {
+		logger.info("Trying to delete account: "+accId);
+		String checkRequest="SELECT COUNT(*) FROM account WHERE account_id=? AND user_id=?";//check if account is customers
+		try {
+			int count =getJdbcTemplate().queryForObject(checkRequest, new Object[] {accId,customerId}, Integer.class);
+			if(count==0) {
+				throw new DataIntegrityViolationException("You can delete only your account!");
+			}
+		} catch (CannotGetJdbcConnectionException e) {
+			logger.error("JDBCConnectionError while deleting account");
+			e.printStackTrace();
+		}catch(DataIntegrityViolationException e) {
+			logger.error("DataIntegrityException while deleting account");
+			e.printStackTrace();
+		}catch(DataAccessException e) {
+			logger.error("DataAccessEx while deleting account");
+			e.printStackTrace();
+		}
+		String deleteQuery = "DELETE FROM account WHERE account_id=? AND user_id=?";
+		try {
+			logger.info("Deleting account: "+accId+ "from database");
+			getJdbcTemplate().update(deleteQuery,new Object[]{accId,customerId});
+		} catch (CannotGetJdbcConnectionException e) {
+			logger.error("JDBCConnectionError while deleting account");
+			e.printStackTrace();
+		}catch(DataAccessException e) {
+			logger.error("DataAccessEx while deleting account");
+			e.printStackTrace();
+		}
 	}
 
 }
